+++ contracts/GSN/IRelayRecipient.sol
@@ -34,16 +34,16 @@ interface IRelayRecipient {
         address relay,
         address from,
         bytes calldata encodedFunction,
-        uint transactionFee,
-        uint gasPrice,
-        uint gasLimit,
-        uint nonce,
+        uint256 transactionFee,
+        uint256 gasPrice,
+        uint256 gasLimit,
+        uint256 nonce,
         bytes calldata approvalData,
-        uint maxPossibleCharge
+        uint256 maxPossibleCharge
     )
         external
         view
-        returns (uint, bytes memory);
+        returns (uint256, bytes memory);
 
     /**
      * @dev Called by {IRelayHub} on approved relay call requests, before the relayed call is executed. This allows to e.g.
@@ -72,5 +72,5 @@ interface IRelayRecipient {
      * and the call to {preRelayedCall} will be reverted retroactively, but the recipient will still be charged for the
      * transaction's cost.
      */
-    function postRelayedCall(bytes calldata context, bool success, uint actualCharge, bytes32 preRetVal) external;
+    function postRelayedCall(bytes calldata context, bool success, uint256 actualCharge, bytes32 preRetVal) external;
 }
